0.86

+++ Booking-Service/src/main/java/com/adamszablewski/service/AppointmentService.java
@@ -1,6 +1,8 @@
 package com.adamszablewski.service;
 
-import com.adamszablewski.dto.mapper.Mapper;
+import com.adamszablewski.dao.Dao;
+import com.adamszablewski.exceptions.NotAuthorizedException;
+import com.adamszablewski.helpers.UserValidator;
 import com.adamszablewski.model.Appointment;
 import com.adamszablewski.helpers.AppointmentHelper;
 import com.adamszablewski.repository.AppointmentRepository;
@@ -10,10 +12,6 @@ import com.adamszablewski.messages.MessageSender;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
-
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import static com.adamszablewski.dto.mapper.Mapper.mapAppointmentToDto;
 
@@ -25,33 +23,44 @@ public class AppointmentService {
     private final AppointmentHelper appointmentHelper;
     private final MessageSender messageSender;
     private final Dao dao;
+    private final UserValidator userValidator;
 
-    public Set<AppoinmentDTO> getAllAppointments() {
-        List<Appointment> appointments = appointmentRepository.findAll();
-        return appointments.stream()
-                .map(Mapper::mapAppointmentToDto)
-                .collect(Collectors.toSet());
-    }
 
-    public AppoinmentDTO getAppointmentById(Long id) {
-        return mapAppointmentToDto(appointmentRepository.findById(id)
-                .orElseThrow(NoSuchAppointmentException::new));
-    }
-    public void deleteAppointmentById(Long id) {
+    public AppoinmentDTO getAppointmentById(Long id, String userEmail) {
         Appointment appointment = appointmentRepository.findById(id)
                 .orElseThrow(NoSuchAppointmentException::new);
+
+        if (!userValidator.isOwnerEmployeeOrTheClient(appointment, userEmail)){
+            throw new NotAuthorizedException();
+        }
+        return mapAppointmentToDto(appointment);
+    }
+    public void deleteAppointmentById(Long id, String userEmail) {
+        Appointment appointment = appointmentRepository.findById(id)
+                .orElseThrow(NoSuchAppointmentException::new);
+        if (!userValidator.isOwner(appointment.getFacility(), userEmail)){
+            throw new NotAuthorizedException();
+        }
         dao.deleteAppoinment(appointment);
         messageSender.createAppoinmentCanceledMessage(appointment);
     }
-    public void changeEmployeeForAppointmentById(Long id) {
+    public void changeEmployeeForAppointmentById(Long id, String userEmail) {
         Appointment appointment= appointmentRepository.findById(id)
                 .orElseThrow(NoSuchAppointmentException::new);
+        if (!userValidator.isOwnerEmployeeOrTheClient(appointment, userEmail)){
+            throw new NotAuthorizedException();
+        }
         appointmentHelper.changeEmployeeForAppointment(appointment, appointment.getEmployee());
     }
 
-    public void markAppointmentAsDone(Long id) {
+    public void markAppointmentAsDone(Long id, String userEmail) {
         Appointment appointment= appointmentRepository.findById(id)
                 .orElseThrow(NoSuchAppointmentException::new);
+
+        if (!userValidator.isEmployee(appointment.getFacility(), userEmail)
+                && !userValidator.isOwner(appointment.getFacility(), userEmail)){
+            throw new NotAuthorizedException();
+        }
         appointmentHelper.deleteAppointment(appointment);
         messageSender.sendAppointmentDoneMessage(appointment);
     }
